Migrate Mutation.deleteToken
